# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    post:
      operationId: createStaffUser
      tags: [admin]
      summary: Create staff user (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"

    get:
      operationId: listStaffUsers
      tags: [admin]
      summary: List all staff users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffUser"
  /admin/audit-log:
    get:
      operationId: getAuditLog
      tags: [admin]
      summary: Query audit log (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: query
          schema:
            type: ["string", "null"]
        - name: actorId
          in: query
          schema:
            type: ["string", "null"]
        - name: action
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StaffUser:
      type: object
      required: [id, email, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [doctor, clinic_admin, nurse]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        mfaEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
    StaffUserInput:
      type: object
      required: [email, password, role, fullName]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 12
        role:
          type: string
          enum: [doctor, clinic_admin, nurse]
        fullName:
          type: string
        phone:
          type: ["string", "null"]

    # ── Patient Auth ────────────────────────────────────────────────────────
...
    AuditLogEntry:
      type: object
      required: [id, action, actorType, outcome, createdAt]
      properties:
        id:
          type: string
        action:
          type: string
        actorType:
          type: string
          enum: [staff, patient, system, unauthenticated]
        actorId:
          type: ["string", "null"]
        actorEmail:
          type: ["string", "null"]
        appointmentId:
          type: ["string", "null"]
        patientId:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        outcome:
          type: string
          enum: [success, failure, blocked]
        detail:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time